requirements.txt
google-generativeai==0.8.6
google-genai==1.60.0

pillow==12.1.0
streamlit==1.53.1
streamlit-option-menu==0.4.0